# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Hospital Vision 2020 Conference 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/set-password:
    post:
      operationId: setPassword
      tags: [auth]
      summary: Set password for first-time login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordInput"
      responses:
        "200":
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
...
  /participants/{participantId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments for a participant
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Add an assignment to a participant
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FILES
...
  /agenda/{registrationNumber}:
    get:
      operationId: getPersonalAgenda
      tags: [participants]
      summary: Get personal agenda by registration number (public)
      parameters:
        - name: registrationNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Personal agenda
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalAgenda"
        "404":
          description: Not found

  # FOOD SESSIONS
...
  /attendance/scan:
    post:
      operationId: scanAttendance
      tags: [attendance]
      summary: Scan QR for attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceScanInput"
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceScanResult"
  /attendance/logs:
    get:
      operationId: listAttendanceLogs
      tags: [attendance]
      summary: List attendance logs
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceLog"

  # DASHBOARD
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [mobile, password]
      properties:
        mobile:
          type: string
        password:
          type: string
        userType:
          type: string
          enum: [participant, admin, track_coordinator, food_coordinator, scientific_committee]
    SetPasswordInput:
      type: object
      required: [mobile, password]
      properties:
        mobile:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
    PasswordResetResponse:
      type: object
      required: [message, resetToken]
      properties:
        message:
          type: string
        resetToken:
          type: string
    ResetPasswordInput:
      type: object
      required: [resetToken, newPassword]
      properties:
        resetToken:
          type: string
        newPassword:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/CurrentUser"
    CurrentUser:
      type: object
      required: [id, name, mobile, userType]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        userType:
          type: string
          enum: [participant, admin, track_coordinator, food_coordinator, scientific_committee]
        assignedTrack:
          type: ["string", "null"]
        participantId:
          type: ["integer", "null"]

    # PARTICIPANT SCHEMAS
    Participant:
      type: object
      required: [id, registrationNumber, name, email, mobile, institution, createdAt]
      properties:
        id:
          type: integer
        registrationNumber:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        institution:
          type: string
        createdAt:
          type: string
          format: date-time
        hasPassword:
          type: boolean
...
    PersonalAgenda:
      type: object
      required: [registrationNumber, name, institution, assignments]
      properties:
        registrationNumber:
          type: string
        name:
          type: string
        institution:
          type: string
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"

    # ASSIGNMENT SCHEMAS
    Assignment:
      type: object
      required: [id, participantId, role, track]
      properties:
        id:
          type: integer
        participantId:
          type: integer
        role:
          type: string
          enum: [Speaker, Presenter, PosterPresenter, Panelist, Moderator, Judge, Chair, CoChair]
        track:
          type: string
        sessionName:
          type: ["string", "null"]
        hall:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        presentationTitle:
          type: ["string", "null"]
        fileId:
          type: ["integer", "null"]
        uploadedFile:
          $ref: "#/components/schemas/UploadedFile"
          nullable: true
    AssignmentInput:
      type: object
      required: [role, track]
      properties:
        role:
          type: string
          enum: [Speaker, Presenter, PosterPresenter, Panelist, Moderator, Judge, Chair, CoChair]
        track:
          type: string
        sessionName:
          type: string
        hall:
          type: string
        date:
          type: string
        time:
          type: string
        presentationTitle:
          type: string
    AssignmentUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [Speaker, Presenter, PosterPresenter, Panelist, Moderator, Judge, Chair, CoChair]
        track:
          type: string
        sessionName:
          type: string
        hall:
          type: string
        date:
          type: string
        time:
          type: string
        presentationTitle:
          type: string

    # FILE SCHEMAS
    UploadedFile:
      type: object
      required: [id, assignmentId, filename, originalName, fileType, version, uploadedAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        fileType:
          type: string
          enum: [pptx, jpg]
        version:
          type: integer
        size:
          type: ["integer", "null"]
        uploadedAt:
          type: string
          format: date-time

    # FOOD SCHEMAS
...
    AttendanceScanInput:
      type: object
      required: [registrationNumber]
      properties:
        registrationNumber:
          type: string
    AttendanceScanResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        participant:
          $ref: "#/components/schemas/Participant"
          nullable: true
    AttendanceLog:
      type: object
      required: [id, participantId, scannedAt]
      properties:
        id:
          type: integer
        participantId:
          type: integer
        participantName:
          type: string
        registrationNumber:
          type: string
        institution:
          type: string
        scannedAt:
          type: string
          format: date-time

    # DASHBOARD SCHEMAS
...